Make swizzles with repeated fields r-values
From the spec:
"Variables that are built-in types, entire structures, structure
fields, l-values with the field selector ( . ) applied to select
components or swizzles without repeated fields, l-values within
parentheses, and l-values dereferenced with the array subscript
operator ( [ ] ) are all l-values. Other unary, binary and ternary
expressions are not l-values. This includes assignments, function
names, swizzles with repeated fields and constants."